backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
python-multipart==0.0.6
openpyxl==3.1.2
pandas==2.1.3
python-dotenv==1.0.0
aiofiles==23.2.1
pydantic==2.5.0